+++ cypress/e2e/test_calculus_graph.cy.js
@@ -24,7 +24,7 @@ describe("Calculus graph", () => {
         cy.get("button[id^='button-id-'][id$='"+jobOne.replaceAll(' ', '-')+"']").should('exist').click();
         cy.get("button[data-bs-target='#collapseCalculatedAttributesWebApplicationJob']").should('exist').click();
 
-
+        cy.get("button[data-bs-target='#collapse-calculated_attributes-ram_needed").should('be.enabled').click();
         cy.get("a[href^='/model_builder/display-calculus-graph/'][href$='ram_needed/']").then(($a) => {
             const url = $a.prop('href');
             cy.visit(url);cy.get('iframe').should('exist');
@@ -60,7 +60,8 @@ describe("Calculus graph", () => {
         cy.get("button[id^='button-id-'][id$='"+jobOne.replaceAll(' ', '-')+"']").should('exist').click();
         cy.get("button[data-bs-target='#collapseCalculatedAttributesWebApplicationJob']").should('exist').click();
 
-        cy.get("a[href^='/model_builder/display-calculus-graph/'][href$='hourly_data_transferred_per_usage_pattern/']").then(($a) => {
+        cy.get("button[data-bs-target='#collapse-calculated_attributes-hourly_occurrences_across_usage_patterns").should('be.enabled').click();
+        cy.get("a[href^='/model_builder/display-calculus-graph/id-4f5352-Test-E2E-Job-1/hourly_occurrences_across_usage_patterns/']").then(($a) => {
             const url = $a.prop('href');
             cy.visit(url);
             cy.get('iframe').should('exist');